Add files via upload
In this commit I have chosen to improve my previous code with improvements on efficiency, automation as well as optimisation. 

1. Created functions for  "Scraping movie details", "Scraping Grossing", "Getting information from TMDB"

2. Added reviews from TMDB, 5 recommended films based on selected films (Link will send you to TMDB page, could send to google(would require using Google Search API))

Improvement: 
User can select a date between 01/01/2021 till to date and it will retrieve the top 15 films of that weekend

User will be able to select a specific film from the top 15 and will get performance data. 

User  can get additional movie information (This will be specified in the front end) 

Reviews will be turned off automatically, user can turn them on if they wish so. 

Future works: 
Compare movies from the same weekend. 

Work on the other information from the spreadsheet with :
1.Rolling 52 week ranking 
2. UK films in top 15
3. UK* share of top 15 gross each weekend
4. UK* share of top 15 overall gross
5: Against last weekend
6. Against same weekend last year
7. Against rolling 52 week norm

May have to find a schedule page for movie releases and find for the following week releasing. If none find the next nearest available job